+++ internal/descriptor/table.go
