+++ internal/descriptor/table_test.go
@@ -3,7 +3,7 @@ package descriptor_test
 import (
 	"testing"
 
-	"github.com/stealthrocket/wasi/wasiunix/internal/descriptor"
+	"github.com/stealthrocket/wasi/internal/descriptor"
 )
 
 type fd uint32
